Version 1.08c
Merged SamNonDuePin library into Arduino core
Implemented rr_config response
Updated SD image directories from RRP's official files (with minor adjustments)
Removed "reset.g" and added support for "stop.g" and "sleep.g"
If M1 is called from the serial interface, M25 is run if Marlin is being emulated
M25 runs "pause.g" only if no live movement is present
M25 runs "resume.g" only if "pause.g" was executed before
G1 S1 sets hit position depending on the homing move's direction
Replaced ENABLE and DISABLE with ENABLE_DRIVE and DISABLE_DRIVE (thanks mike-o-tronic)
G-Code response buffer is only cleared when it is actually polled (to view start-up errors from the web interface)
Bug fix: Firmware didn't report a message at all if no SD card was present
Bug fix: M20 returns proper file list if Marlin is emulated
Bug fix: Only temperatures less than -273C switch off all heaters (to retain compatibility with RRP's config files)
Bug fix: G32 occasionally caused a watchdog reset if bed.g was used
Other minor changes

+++ Move.cpp
@@ -1180,7 +1180,7 @@ void Move::SetProbedBedEquation(StringRef& reply)
 	}
 
 	reply.copy("Bed equation fits points");
-	for(int8_t point = 0; point < NumberOfProbePoints(); point++)
+	for(size_t point = 0; point < NumberOfProbePoints(); point++)
 	{
 		reply.catf(" [%.1f, %.1f, %.3f]", xBedProbePoints[point], yBedProbePoints[point], zBedProbePoints[point]);
 	}
